+++ packages/leaflet-searchaddress/src/index.ts
@@ -1,6 +1,11 @@
 import { SearchAddressControl } from "./SearchAddressControl";
 import { SearchAddress, SearchAddressOptions } from "./SearchAddress";
 import { SearchPositionControl, SearchPositionOptions } from "./SearchPositionControl";
+import { Address } from "./Address";    
+
+if (typeof window != "undefined") {
+    (window as any).Address = Address;
+}
 
 // UMD globals (only if Leaflet is loaded)
 if (typeof window !== "undefined" && (window as any).L) {
@@ -17,5 +22,7 @@ if (typeof window !== "undefined" && (window as any).L) {
 
 }
 
+
+export { Address }
 export { SearchAddressControl, SearchAddress, SearchPositionControl };
 export type { SearchAddressOptions, SearchPositionOptions };