+++ packages/shared/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @locatienet/shared
 
+## 1.0.4
+
+### Patch Changes
+
+- Automated version bump
+
+## 1.0.3
+
+### Patch Changes
+
+- Automated version bump
+
+## 1.0.2
+
+### Patch Changes
+
+- Automated version bump
+
 ## 1.0.1
 
 ### Patch Changes